pyproject.toml
[tool.poetry]
name = "testbotm"
version = "0.1.0"
description = ""
authors = ["polenom <vitalimitsevich@yandex.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyrogram = "^2.0.106"
python-dotenv = "^1.0.1"
tgcrypto = "^1.2.5"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
